pyramid-rag/backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
email-validator==2.2.0

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
Pillow==11.0.0
python-magic==0.4.27
chardet==3.0.4
textract==1.6.5
cairosvg==2.7.1
ezdxf==1.3.4

# Embeddings & ML
sentence-transformers==3.2.1
torch==2.5.1
transformers==4.46.2
langchain==0.3.7
langchain-community==0.3.5
chromadb==0.5.18
tiktoken==0.8.0

# Task Queue
celery[redis]==5.4.0
redis==5.2.0
flower==2.0.1

# API Utils
httpx==0.27.2
aiofiles==24.1.0
python-dateutil==2.9.0

# Monitoring
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
psutil==5.9.8

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==30.8.1

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
pre-commit==4.0.1